+++ lib/screen/track_order/widgets/delivery_time_address.dart
@@ -1,0 +1,23 @@
+import 'package:himalayan_delights/screen/track_order/imports.dart';
+
+class DeliveryTimeAddress extends StatelessWidget {
+  const DeliveryTimeAddress({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    const EdgeInsetsGeometry gapPadding = EdgeInsets.symmetric(horizontal: 15);
+
+    return const Padding(
+      padding: gapPadding,
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        children: [
+          DeliveryTime(),
+          DeliveryAddress(),
+        ],
+      ),
+    );
+  }
+}
